Moved tests to the right directory
